dr-plan-generator/requirements.txt
boto3>=1.34
requests>=2.28
pyyaml>=6.0
